p/router fixed for edit & create then view and common hook aplay for create & edit
